pyproject.toml
[project]
name = "ai-agent-for-inbound-carrier-sales"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ace-tools>=0.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.2",
    "langsmith>=0.4.10",
    "openai>=1.98.0",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
]
